backend/requirements.txt
# Web Framework
fastapi>=0.111.0
uvicorn[standard]>=0.30.0

# Database
sqlmodel==0.0.14
psycopg2-binary>=2.9.9
alembic>=1.12.1

# Authentication
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt>=4.0.0
python-multipart>=0.0.9

# Validation
pydantic[email]>=2.7.0
pydantic-settings>=2.2.0
email-validator>=2.1.0

# Environment
python-dotenv>=1.0.0

# Testing
pytest>=7.4.3
pytest-asyncio>=0.23.0
httpx>=0.27.0
pytest-cov>=4.1.0

# AI Integration
openai>=1.50.0
mcp>=1.0.0

# Code Quality
ruff>=0.1.6
black>=23.11.0
PyJWT>=2.8.0
